refactor: 상수용 패키지 생성
