+++ src/main/java/com/coupang/marketplace/controller/SignInRequestDto.java
@@ -2,7 +2,6 @@ package com.coupang.marketplace.controller;
 
 import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
